fix: non-zero delay between first attempt and first retry for linear and exp strategy (#163)
* tests: static backoff: check delay

See issue #122.

This is to confirm that the delay
between the first (actual) attempt
and the first retry is not zero,
but the configured delay.

tests: prettier format (noop)

tests: prettier format (noop)

* fix delay calc: non-zero between first two attempts

See issue #122.

* calc delay: fix currentRetryAttempt mutation

The mutation did not affect the code below.
This is for issue #122.

calc delay: prettier (format)

* tests: test delay for linear, exp strategies

* calc delay: mutate cfg on orig err obj

Found that state is not retained across
retries when mutating
`config.currentRetryAttempt` here.

* calc delay: use expressive variable name

* package.json: require nodejs 10.7.0 or newer

This is to address the `gts fix` error

    The 'process.hrtime.bigint' is not supported
    until Node.js 10.7.0. The configured version
    range is '>=10.0.0'

This function is only used in the test suite
so maybe it's not the best solution to change
the NodeJS version requirement for the
entire package. However, whoever runs this on
NodeJS older than 10.7.0 (released on
2018-07-18) might want to be encouraged to
use a more recent version.